fix my-offering page
