+++ frontend/src/components/ProducerWorkingContent.tsx
@@ -1,5 +1,5 @@
 import { Calendar } from "lucide-react";
-import ProductionWorkingCard from "./ProductionWorkingCard";
+import ProductionWorkingCard from "./ProducerWorkingCard";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "./ui/select";
 import { useState } from "react";
 import { Card } from "./ui/card";
@@ -37,7 +37,7 @@ const offers = [
     }
 ];
 
-export default function ProductionWorkingContent(){
+export default function ProducerWorkingContent(){
     const [selectedRole, setSelectedRole] = useState("Actor");
     const [selectedProfessional, setSelectedProfessional] = useState("Tendo Souji");
 
@@ -86,45 +86,4 @@ export default function ProductionWorkingContent(){
         </div>
         </main>
       );
-    {/*return(
-        <main className="bg-slate-480 rounded-lg h-full shadow-xl m-auto w-[100%]">
-            <div className="grid grid-cols-1 h-full w-f">
-                <h1 className="text-start text-xl font-bold my-5 ml-10 p-0">Offer History</h1>
-                <div className="w-[90%] m-auto content-center">
-                    <div className="w-full grid grid-cols-11 h-full bg-mainblue-darkest m-auto text-sm p-5 font-bold text-white text-center whitespace-normal break-words items-center content-center rounded-lg">
-                        <div className="col-span-2">
-                            <p className="">Name</p>
-                        </div>
-                        <div className="col-span-2">
-                            <p className="">Role</p>
-                        </div>
-                        <div className="col-span-2">
-                            <p className="">Status</p>
-                        </div>
-                        <div className="col-span-3">
-                            <p className="">Confirmation</p>
-                        </div>
-                        <div className="col-span-2">
-                            <p className="">Work Quota</p>
-                        </div>
-                    </div>
-                    <div className="overflow-y-auto h-full max-h-[450px] m-auto bg-slate-400">
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                        <ProductionWorkingCard/>
-                    </div>
-                    
-                </div>
-            </div>
-        </main>
-    )*/}
 }
